pyproject.toml
[tool.poetry]
name = "fastapitemplate"
version = "0.1.0"
description = "FastAPI async template"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
black = "^22.3.0"
isort = "^5.10.1"
uvicorn = "^0.17.6"
greenlet = "^1.1.2"
httpx = "^0.23.0"
fastapi = "^0.87.0"
sqlalchemy = "^2.0.0b3"
alembic = "^1.7.7"
python-dotenv = "^0.20.0"
pydantic = "^1.10.2"
asyncpg = "^0.25.0"
psycopg2 = "^2.9.3"
passlib = "^1.7.4"
email-validator = "^1.2.1"
bcrypt = "^3.2.2"
python-dateutil = "^2.8.2"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
